WIP - Split gulp tasks. Ref #130.
First go a breaking up gulp tasks.
Got some issues.
